# src/main/resources/com/sonicle/webtop/contacts/openapi-v2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebTop Contacts",
    "contact": {
      "email": "dev-team@sonicle.com"
    },
    "version": "v2",
    "description": "This is Contacts service API enpoint."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "auth-basic": []
    },
    {
      "auth-bearer": []
    }
  ],
  "paths": {
    "/carddav/addressbooks/{addressBookUid}": {
      "get": {
        "tags": [
          "dav"
        ],
        "summary": "Gets a single address-book",
        "operationId": "getAddressBook",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DavAddressBook"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          },
          "404": {
            "description": "Address book not found",
            "content": {}
          }
        },
        "description": "Gets the specified AddressBook."
      },
      "put": {
        "tags": [
          "dav"
        ],
        "summary": "Updates an address-book",
        "operationId": "updateAddressBook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DavAddressBookUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          },
          "404": {
            "description": "Address book not found",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body",
        "description": "Updates specified AddressBook."
      },
      "delete": {
        "tags": [
          "dav"
        ],
        "summary": "Deletes an address-book",
        "operationId": "deleteAddressBook",
        "responses": {
          "204": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          },
          "404": {
            "description": "Address book not found",
            "content": {}
          },
          "405": {
            "description": "Delete operation is not allowed",
            "content": {}
          }
        },
        "description": "Deletes specified AddressBook."
      },
      "parameters": [
        {
          "name": "addressBookUid",
          "in": "path",
          "description": "Address book UID",
          "required": true,
          "schema": {
            "type": "string"
          },
          "example": "1234"
        }
      ]
    },
    "/carddav/addressbooks": {
      "get": {
        "tags": [
          "dav"
        ],
        "summary": "List all address-books",
        "operationId": "getAddressBooks",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DavAddressBook"
                  }
                }
              }
            }
          }
        },
        "description": "List available AddressBooks."
      },
      "post": {
        "tags": [
          "dav"
        ],
        "summary": "Adds a new address-book",
        "operationId": "addAddressBook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DavAddressBookNew"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DavAddressBook"
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "body",
        "description": "Creates new AddressBook."
      }
    },
    "/carddav/addressbooks/{addressBookUid}/cards": {
      "get": {
        "tags": [
          "dav"
        ],
        "summary": "List all cards for a specific address-book",
        "operationId": "getCards",
        "parameters": [
          {
            "name": "hrefs",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DavCard"
                  }
                }
              }
            }
          }
        },
        "description": "List all Cards of specified AddressBook."
      },
      "post": {
        "tags": [
          "dav"
        ],
        "summary": "Adds a new card",
        "operationId": "addCard",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DavCardNew"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body",
        "description": "Creates new Card into specified AddressBook."
      },
      "parameters": [
        {
          "name": "addressBookUid",
          "in": "path",
          "description": "Address book UID",
          "required": true,
          "schema": {
            "type": "string"
          },
          "example": "1234"
        }
      ]
    },
    "/carddav/addressbooks/{addressBookUid}/cards/changes": {
      "get": {
        "tags": [
          "dav"
        ],
        "summary": "Get card changes",
        "description": "Returns changed cards (added/modified/deleted) since the specified syncToken. If token is not provided, the initial sync configuration will be returned.",
        "operationId": "getCardsChanges",
        "parameters": [
          {
            "name": "syncToken",
            "in": "query",
            "description": "Marks changes starting point",
            "schema": {
              "type": "string"
            },
            "example": "20180416130954"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limits the number of returned results",
            "schema": {
              "type": "integer"
            },
            "example": 100
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DavCardsChanges"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          }
        }
      },
      "parameters": [
        {
          "name": "addressBookUid",
          "in": "path",
          "description": "Address book UID",
          "required": true,
          "schema": {
            "type": "string"
          },
          "example": "1234"
        }
      ]
    },
    "/carddav/addressbooks/{addressBookUid}/cards/{href}": {
      "get": {
        "tags": [
          "dav"
        ],
        "summary": "Get a single card",
        "operationId": "getCard",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DavCard"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          },
          "404": {
            "description": "Card not found",
            "content": {}
          }
        },
        "description": "Gets specified Card."
      },
      "put": {
        "tags": [
          "dav"
        ],
        "summary": "Updates a card",
        "operationId": "updateCard",
        "requestBody": {
          "content": {
            "text/vcard": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          },
          "404": {
            "description": "Card not found",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body",
        "description": "Updates specified Card."
      },
      "delete": {
        "tags": [
          "dav"
        ],
        "summary": "Deletes a card",
        "operationId": "deleteCard",
        "responses": {
          "204": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          },
          "404": {
            "description": "Card not found",
            "content": {}
          }
        },
        "description": "Deletes specified Card."
      },
      "parameters": [
        {
          "name": "addressBookUid",
          "in": "path",
          "description": "Address book UID",
          "required": true,
          "schema": {
            "type": "string"
          },
          "example": "1234"
        },
        {
          "name": "href",
          "in": "path",
          "description": "Card reference URI",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
...
  },
  "components": {
    "schemas": {
...
      "DavAddressBook": {
        "type": "object",
        "description": "Bean for carry addressbook's fields",
        "properties": {
          "id": {
            "type": "string",
            "description": "AddressBook ID (internal)"
          },
          "uid": {
            "type": "string",
            "description": "AddressBook UID (public)"
          },
          "displayName": {
            "type": "string",
            "description": "Display name"
          },
          "description": {
            "type": "string",
            "description": "Description"
          },
          "syncToken": {
            "type": "string",
            "description": "Current sync token"
          },
          "aclFol": {
            "type": "string",
            "description": "ACL info for folder itself"
          },
          "aclEle": {
            "type": "string",
            "description": "ACL info for folder elements"
          },
          "ownerUsername": {
            "type": "string",
            "description": "The owner profile's username"
          }
        },
        "required": [
          "id",
          "uid",
          "displayName",
          "syncToken",
          "aclFol",
          "aclEle",
          "ownerUsername"
        ]
      },
      "DavAddressBookUpdate": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "description": "New value for displayName"
          },
          "description": {
            "type": "string",
            "description": "New value for description"
          },
          "updatedFields": {
            "type": "array",
            "description": "Specifies which fields have been updated",
            "items": {
              "type": "string"
            }
          }
        },
        "description": "Bean for carry addressbook's updateable fields"
      },
      "DavAddressBookNew": {
        "required": [
          "displayName"
        ],
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "description": "Display name"
          },
          "description": {
            "type": "string",
            "description": "Description"
          }
        },
        "description": "Bean for carry addressbook's fields"
      },
      "DavCard": {
        "type": "object",
        "description": "Bean for carry card fields",
        "properties": {
          "id": {
            "type": "string",
            "description": "Card ID (internal)"
          },
          "uid": {
            "type": "string",
            "description": "Card UID (public)"
          },
          "href": {
            "type": "string",
            "description": "Reference URI"
          },
          "lastModified": {
            "type": "integer",
            "description": "Last modification time (unix timestamp)",
            "format": "int64"
          },
          "etag": {
            "type": "string",
            "description": "Revision tag"
          },
          "size": {
            "type": "integer",
            "description": "Size (in bytes) of card data"
          },
          "vcard": {
            "type": "string",
            "description": "Card data (vCard format)"
          }
        },
        "required": [
          "id",
          "uid",
          "href"
        ]
      },
      "DavCardNew": {
        "required": [
          "href",
          "vcard"
        ],
        "type": "object",
        "properties": {
          "href": {
            "type": "string",
            "description": "Reference URI"
          },
          "vcard": {
            "type": "string",
            "description": "Card data (vCard format)"
          }
        },
        "description": "Bean for carry new card fields"
      },
      "DavCardChanged": {
        "type": "object",
        "description": "Bean for carry fields of card that have been changed",
        "properties": {
          "id": {
            "type": "string",
            "description": "Card ID (internal)"
          },
          "href": {
            "type": "string",
            "description": "Reference URI"
          },
          "etag": {
            "type": "string",
            "description": "Revision tag"
          }
        },
        "required": [
          "id",
          "href",
          "etag"
        ]
      },
      "DavCardsChanges": {
        "required": [
          "deleted",
          "inserted",
          "syncToken",
          "updated"
        ],
        "type": "object",
        "properties": {
          "syncToken": {
            "type": "string",
            "description": "Current sync token"
          },
          "inserted": {
            "type": "array",
            "description": "Items that have been inserted",
            "items": {
              "$ref": "#/components/schemas/DavCardChanged"
            }
          },
          "updated": {
            "type": "array",
            "description": "Items that have been updated",
            "items": {
              "$ref": "#/components/schemas/DavCardChanged"
            }
          },
          "deleted": {
            "type": "array",
            "description": "Items that have been deleted",
            "items": {
              "$ref": "#/components/schemas/DavCardChanged"
            }
          }
        },
        "description": "Bean for carry card collection changes"
      },
...
    },
    "securitySchemes": {
      "auth-basic": {
        "type": "http",
        "scheme": "basic"
      },
      "auth-bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
...
}